wip: chore: allow partial parse trees
This allows `Parser.Parse` to return a partial parse tree - the parser
is permitted to return an incomplete parse tree alongside the error.

This is useful for syntax highlighting, auto-completion, and the like:
in these scenarios, we want to be able to access a partial *best-effort*
parse tree, even if the parsing has failed.

Signed-off-by: Justin Chadwell <me@jedevc.com>