wip: hack: avoid walking nil elements
In a well-structured parse tree, this won't happen. But if we have a
partial parse tree, as added in the previous commit, we need some way to
avoid panics while traversing.

This requires checking `node == nil` at the top of the `Walk` function,
as well as adding a little special generic magic to avoid accidentally
passing a *typed nil* into the `Walk` function (which isn't detected by
that check).

(yes there are better ways of doing this, it's a wip hack)

Signed-off-by: Justin Chadwell <me@jedevc.com>